# docs/api-docs.yaml
openapi: 3.0.3
info:
  title: Registry Viewer API
  description: API for registry viewer to obtain registry data
  contact:
    email: myung.choi@gtri.gatech.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: v1.11.1
servers:
  - url: https://smartchartsuite.dev.heat.icl.gtri.org
    description: Production server
  - url: http://localhost:8080
    description: Local development (optional)
...
paths:
  /registry-viewer-api/case-record/{registry}:
    get:
      tags:
        - case-record-api-controller
      summary: searches context within a category
      description: "By passing in the appropriate options, you can search for available cases in the registry "
      operationId: searchCategoryUsingGET
      parameters:
        - name: registry
          in: path
          description: Registry Path
          required: true
          style: simple
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]+$"
        - name: caseId
          in: query
          description: case-id for the category
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
        - name: sections
          in: query
          description: sections to query for the case-id
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
            description: CSV list of section identifiers
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseData'
        '400':
          description: Bad request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - Authorization: []
        - basicAuth: []
    put:
      tags:
        - case-record-api-controller
      summary: Add a flag, annotation, or user data
      description: Add a flag, annotation, or user data
      operationId: addUserFlagAnnotationManualDataUsingPUT
      parameters:
        - name: registry
          in: path
          description: Registry Path
          required: true
          style: simple
          schema:
            type: string
        - name: caseId
          in: query
          description: caseId
          required: true
          style: form
          schema:
            type: integer
            format: int32
        - name: contentId
          in: query
          description: contentId (optional, identifies content to update)
          required: false
          style: form
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserFlagAnnotationManualData'
      responses:
        '200':
          description: requested data updated
        '201':
          description: requested data created
        '400':
          description: invalid input, object invalid
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - Authorization: []
        - basicAuth: []
...
components:
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header
      description: API key passed in the `Authorization` header (e.g., `Bearer <token>`)
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication (optional)
  schemas:
    Error:
      type: object
      description: Generic error response
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: Invalid request parameters
      required:
        - code
        - message
    Annotation:
      title: Annotation
      type: object
      properties:
        annotationId:
          type: integer
          format: int32
          example: 1
        date:
          type: string
          format: date-time
          example: '2024-01-04T05:00:00Z'
        text:
          type: string
          example: will be deleted if annotationId is valid but text is empty
    CaseData:
      title: CaseData
      type: object
      properties:
        caseId:
          type: integer
          format: int32
          example: 1
        contents:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        count:
          type: integer
          format: int32
          example: 1
...
    Coding:
      title: Coding
      type: object
      properties:
        code:
          type: string
          example: 20507-0
        display:
          type: string
          example: Reagin Ab [Presence] in Serum by RPR
        system:
          type: string
          example: LOINC
    Content:
      title: Content
      type: object
      properties:
        annotation:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        category:
          type: string
          example: Syphilis
        contentId:
          type: integer
          format: int32
          example: 1
        date:
          type: string
          format: date-time
          example: '2022-01-14T05:00:00Z'
        derivedValue:
          $ref: '#/components/schemas/Value'
        details:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/DetailCondition'
              - $ref: '#/components/schemas/DetailMedication'
              - $ref: '#/components/schemas/DetailObservation'
              - $ref: '#/components/schemas/DetailMeasurement'
              - $ref: '#/components/schemas/DetailNote'
              - $ref: '#/components/schemas/DetailUserData'
        flag:
          type: string
          example: flag1
        question:
          type: string
          example: Lab Result
        section:
          type: string
          example: LabResults
        sourceValue:
          $ref: '#/components/schemas/Value'
    DetailBase:
      type: object
      properties:
        code:
          type: string
          example: 20507-0
        display:
          type: string
          example: Reagin Ab [Presence] in Serum by RPR
        startDate:
          type: string
          format: date-time
          example: '2022-01-14T05:00:00Z'
        endDate:
          type: string
          format: date-time
          example: '2022-01-14T05:00:00Z'
        system:
          type: string
          example: LOINC
        tableDisplayText:
          type: string
          example: Reagin Ab [Presence] in Serum by RPR | value
    DetailCondition:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
    DetailMeasurement:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
        - type: object
          properties:
            date:
              type: string
              format: date-time
              example: '2022-01-14T05:00:00Z'
            rangeHigh:
              type: integer
              format: int32
              example: 5
            rangeLow:
              type: integer
              format: int32
              example: 1
            unit:
              type: string
            value:
              type: string
              example: Reactive
    DetailMedication:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
        - type: object
          properties:
            daysSupply:
              type: integer
              format: int32
              example: 30
            lotNumber:
              type: string
              example: lot number
            quantity:
              type: integer
              format: int32
              example: 5
            refills:
              type: integer
              format: int32
              example: 2
            route_code:
              type: string
              example: 1234-5
            route_display:
              type: string
              example: route text info
            route_system:
              type: string
            sig:
              type: string
            startDate:
              type: string
              format: date-time
              example: '2022-01-14T05:00:00Z'
    DetailNote:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
        - type: object
          properties:
            noteText:
              type: string
              example: Reactive
    DetailObservation:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
        - type: object
          properties:
            unit:
              type: string
            value:
              type: string
              example: Reactive
    DetailUserData:
      allOf:
        - $ref: '#/components/schemas/DetailBase'
...
    ManualCaseData:
      title: ManualCaseData
      type: object
      properties:
        conceptId:
          type: integer
          format: int32
          example: 1
        date:
          type: string
          format: date-time
          example: '2024-01-04T05:00:00Z'
        value:
          type: string
          example: user-entered value
...
    UserFlagAnnotationManualData:
      title: UserFlagAnnotationManualData
      type: object
      properties:
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        flag:
          type: string
          example: Invalid Entry
        manualCaseData:
          type: array
          items:
            $ref: '#/components/schemas/ManualCaseData'
    Value:
      title: Value
      type: object
      properties:
        coding:
          $ref: '#/components/schemas/Coding'
        unit:
          type: string
          example: unit if value is numeric
        value:
          type: string
          example: Reactive
...